# Push down query filter to Parquet conversion.