--- a/query/physicalplan/filter.go
+++ b/query/physicalplan/filter.go
@@ -11,6 +11,7 @@ import (
 	"github.com/apache/arrow/go/v15/arrow/array"
 	"github.com/apache/arrow/go/v15/arrow/memory"
 	"github.com/apache/arrow/go/v15/arrow/scalar"
+	"github.com/parquet-go/parquet-go"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/polarsignals/frostdb/query/logicalplan"
@@ -40,6 +41,7 @@ func NewBitmap() *Bitmap {
 
 type BooleanExpression interface {
 	Eval(r arrow.Record) (*Bitmap, error)
+	EvalParquet(rg parquet.RowGroup, in [][]parquet.Value) (*Bitmap, [][]parquet.Value, error)
 	String() string
 }
 
@@ -130,12 +132,12 @@ func binaryBooleanExpr(expr *logicalplan.BinaryExpr) (BooleanExpression, error) 
 			Right: rightScalar,
 		}, nil
 	case logicalplan.OpAnd:
-		left, err := booleanExpr(expr.Left)
+		left, err := BooleanExpr(expr.Left)
 		if err != nil {
 			return nil, fmt.Errorf("left bool expr: %w", err)
 		}
 
-		right, err := booleanExpr(expr.Right)
+		right, err := BooleanExpr(expr.Right)
 		if err != nil {
 			return nil, fmt.Errorf("right bool expr: %w", err)
 		}
@@ -145,12 +147,12 @@ func binaryBooleanExpr(expr *logicalplan.BinaryExpr) (BooleanExpression, error) 
 			Right: right,
 		}, nil
 	case logicalplan.OpOr:
-		left, err := booleanExpr(expr.Left)
+		left, err := BooleanExpr(expr.Left)
 		if err != nil {
 			return nil, fmt.Errorf("left bool expr: %w", err)
 		}
 
-		right, err := booleanExpr(expr.Right)
+		right, err := BooleanExpr(expr.Right)
 		if err != nil {
 			return nil, fmt.Errorf("right bool expr: %w", err)
 		}
@@ -167,6 +169,26 @@ func binaryBooleanExpr(expr *logicalplan.BinaryExpr) (BooleanExpression, error) 
 type AndExpr struct {
 	Left  BooleanExpression
 	Right BooleanExpression
+}
+
+func (a *AndExpr) EvalParquet(rg parquet.RowGroup, in [][]parquet.Value) (*Bitmap, [][]parquet.Value, error) {
+	left, lout, err := a.Left.EvalParquet(rg, in)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	if left.IsEmpty() {
+		return left, nil, nil
+	}
+
+	right, rout, err := a.Right.EvalParquet(rg, lout)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// This stores the result in place to avoid allocations.
+	left.And(right)
+	return left, rout, nil
 }
 
 func (a *AndExpr) Eval(r arrow.Record) (*Bitmap, error) {
@@ -198,6 +220,22 @@ type OrExpr struct {
 	Right BooleanExpression
 }
 
+func (a *OrExpr) EvalParquet(rg parquet.RowGroup, in [][]parquet.Value) (*Bitmap, [][]parquet.Value, error) {
+	left, lout, err := a.Left.EvalParquet(rg, in)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	right, rout, err := a.Right.EvalParquet(rg, lout)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// This stores the result in place to avoid allocations.
+	left.Or(right)
+	return left, rout, nil
+}
+
 func (a *OrExpr) Eval(r arrow.Record) (*Bitmap, error) {
 	left, err := a.Left.Eval(r)
 	if err != nil {
@@ -218,17 +256,22 @@ func (a *OrExpr) String() string {
 	return "(" + a.Left.String() + " OR " + a.Right.String() + ")"
 }
 
-func booleanExpr(expr logicalplan.Expr) (BooleanExpression, error) {
+func BooleanExpr(expr logicalplan.Expr) (BooleanExpression, error) {
 	switch e := expr.(type) {
 	case *logicalplan.BinaryExpr:
 		return binaryBooleanExpr(e)
+	case *logicalplan.AggregationFunction:
+		// This happens in the case of an aggregation function with no group by clause.
+		// It gets pushed down into the filter layer, but this isn't supported as a physical
+		// boolean expression yet.
+		return nil, nil
 	default:
 		return nil, ErrUnsupportedBooleanExpression
 	}
 }
 
 func Filter(pool memory.Allocator, tracer trace.Tracer, filterExpr logicalplan.Expr) (*PredicateFilter, error) {
-	expr, err := booleanExpr(filterExpr)
+	expr, err := BooleanExpr(filterExpr)
 	if err != nil {
 		return nil, fmt.Errorf("create bool expr: %w", err)
 	}