Push the query filter down to the Parquet row group conversion layer. This saves on CPU time when building Arrow records since our row group usage tends to be sparse.

```
name               old time/op    new time/op    delta
Query/MergeAll-10     1.34s ± 1%     0.96s ± 1%  -28.18%  (p=0.000 n=10+10)
Query/Merge-10        1.33s ± 2%     0.96s ± 1%  -28.27%  (p=0.000 n=10+9)

name               old alloc/op   new alloc/op   delta
Query/MergeAll-10    5.51GB ± 0%    5.64GB ± 1%   +2.33%  (p=0.000 n=10+9)
Query/Merge-10       5.52GB ± 1%    5.65GB ± 1%   +2.35%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
Query/MergeAll-10     34.1M ± 0%     50.6M ± 0%  +48.35%  (p=0.000 n=10+10)
Query/Merge-10        34.1M ± 0%     50.6M ± 0%  +48.34%  (p=0.000 n=10+9)
```