@thorfour 2024-03-26
Latest benchmarks

```bash
thor@thors-MacBook-Pro ~/.../github.com/polarsignals/frostdb % benchstat before.txt latest.txt
name               old time/op    new time/op    delta
Query/MergeAll-10     1.34s ± 1%     0.85s ± 1%  -36.38%  (p=0.000 n=10+8)
Query/Merge-10        1.33s ± 2%     0.85s ± 2%  -35.93%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
Query/MergeAll-10    5.51GB ± 0%    4.55GB ± 1%  -17.52%  (p=0.000 n=10+10)
Query/Merge-10       5.52GB ± 1%    4.55GB ± 1%  -17.62%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
Query/MergeAll-10     34.1M ± 0%     47.6M ± 0%  +39.49%  (p=0.000 n=10+10)
Query/Merge-10        34.1M ± 0%     47.6M ± 0%  +39.49%  (p=0.000 n=10+10)
```